Enhancement/post api

+++ Letterbook.Core/PostService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Runtime.CompilerServices;
+using System.Security.Claims;
 using Letterbook.Core.Adapters;
 using Letterbook.Core.Exceptions;
 using Letterbook.Core.Extensions;
@@ -14,13 +15,15 @@ namespace Letterbook.Core;
 /// <summary>
 /// The PostService is how you Post™
 /// </summary>
-public class PostService : IPostService
+public class PostService : IAuthzPostService, IPostService
 {
     private readonly ILogger<PostService> _logger;
     private readonly CoreOptions _options;
     private readonly IPostAdapter _posts;
     private readonly IPostEventService _postEvents;
     private readonly IActivityPubClient _apClient;
+    private Uuid7 _profileId;
+    private IEnumerable<Claim> _claims = null!;
 
     public PostService(ILogger<PostService> logger, IOptions<CoreOptions> options,
         IPostAdapter posts, IPostEventService postEvents, IActivityPubClient apClient)
@@ -75,7 +78,7 @@ public class PostService : IPostService
         return await Draft(post, inReplyToId);
     }
 
-    public async Task<Post> Draft(Post post, Uuid7? inReplyToId = default)
+    public async Task<Post> Draft(Post post, Uuid7? inReplyToId = default, bool publish = false)
     {
         if (inReplyToId is { } parentId)
         {
@@ -87,38 +90,48 @@ public class PostService : IPostService
             post.Thread.Posts.Add(post);
         }
 
+        if (await _posts.LookupProfile(_profileId) is not { } author)
+	        throw CoreException.MissingData<Profile>($"Couldn't find profile {_profileId}", _profileId);
+        post.Creators.Clear();
+        post.Creators.Add(author);
+
+        if (publish) post.PublishedDate = DateTimeOffset.UtcNow;
         _posts.Add(post);
         await _posts.Commit();
         _postEvents.Created(post);
-        
+        if (publish) _postEvents.Published(post);
+
         return post;
     }
 
-    public async Task<Post> Update(Post post)
+    public async Task<Post> Update(Uuid7 postId, Post post)
     {
         // TODO: authz
         // I think authz can be conveyed around the app with just a list of claims, as long as one of the claims is
         // a profile, right?
-        var previous = await _posts.LookupPost(post.Id)
+        var previous = await _posts.LookupPost(postId)
                        ?? throw CoreException.MissingData($"Could not find existing post {post.Id} to update",
-                           typeof(Post), post.Id);
-        
+                           typeof(Post), postId);
+
+        if (!(previous as IFederated).StrictEqual(post))
+	        throw CoreException.InvalidRequest("Input IDs don't match", $"{postId.ToId25String()}", post);
+        // var decision = _authz.Update(Enumerable.Empty<Claim>(), previous) // TODO: authz
         previous.Client = post.Client; // probably should come from an authz claim
         previous.InReplyTo = post.InReplyTo;
         previous.Audience = post.Audience;
-        
+
         // remove all the removed contents, and add/update everything else
         var removed = previous.Contents.Except(post.Contents).ToArray();
         _posts.RemoveRange(removed);
         previous.Contents = post.Contents;
-        
+
         var published = previous.PublishedDate != null;
         if (published)
         {
-            previous.UpdatedDate = DateTimeOffset.Now;
+            previous.UpdatedDate = DateTimeOffset.UtcNow;
             // publish again, tbd
         }
-        else previous.CreatedDate = DateTimeOffset.Now;
+        else previous.CreatedDate = DateTimeOffset.UtcNow;
 
 
         _posts.Update(previous);
@@ -133,7 +146,7 @@ public class PostService : IPostService
         var post = await _posts.LookupPost(id);
         if (post is null) return;
         // TODO: authz and thread root
-        post.DeletedDate = DateTimeOffset.Now;
+        post.DeletedDate = DateTimeOffset.UtcNow;
         _posts.Remove(post);
         await _posts.Commit();
         _postEvents.Deleted(post);
@@ -145,9 +158,9 @@ public class PostService : IPostService
         if (post is null) throw CoreException.MissingData<Post>($"Can't find post {id} to publish", id);
         if (post.PublishedDate is not null)
             throw CoreException.Duplicate($"Tried to publish post {id} that is already published", id);
-        post.PublishedDate = DateTimeOffset.Now;
-        post.CreatedDate = DateTimeOffset.Now;
-        
+        post.PublishedDate = DateTimeOffset.UtcNow;
+        post.CreatedDate = DateTimeOffset.UtcNow;
+
         _posts.Update(post);
         await _posts.Commit();
         _postEvents.Published(post);
@@ -156,18 +169,19 @@ public class PostService : IPostService
 
     public async Task<Post> AddContent(Uuid7 postId, Content content)
     {
-        var post = await _posts.LookupPost(postId) 
+        var post = await _posts.LookupPost(postId)
                    ?? throw CoreException.MissingData<Post>("Can't find existing post to add content", postId);
         if (!post.FediId.HasLocalAuthority(_options))
             throw CoreException.WrongAuthority("Can't modify contents of remote post", post.FediId);
+        content.SortKey ??= (post.Contents.Select(c => c.SortKey).Max() ?? -1) + 1;
         post.Contents.Add(content);
-        
+
         if (post.PublishedDate is not null)
         {
-            post.UpdatedDate = DateTimeOffset.Now;
+            post.UpdatedDate = DateTimeOffset.UtcNow;
             _postEvents.Published(post);
         }
-        
+
         _posts.Update(post);
         await _posts.Commit();
         _postEvents.Updated(post);
@@ -176,7 +190,7 @@ public class PostService : IPostService
 
     public async Task<Post> RemoveContent(Uuid7 postId, Uuid7 contentId)
     {
-        var post = await _posts.LookupPost(postId) 
+        var post = await _posts.LookupPost(postId)
                    ?? throw CoreException.MissingData<Post>("Can't find existing post to remove content", postId);
         if (!post.FediId.HasLocalAuthority(_options))
             throw CoreException.WrongAuthority("Can't modify contents of remote post", post.FediId);
@@ -188,13 +202,13 @@ public class PostService : IPostService
         }
 
         post.Contents.Remove(content);
-        
+
         if (post.PublishedDate is not null)
         {
-            post.UpdatedDate = DateTimeOffset.Now;
+            post.UpdatedDate = DateTimeOffset.UtcNow;
             _postEvents.Published(post);
         }
-        
+
         _posts.Remove(content);
         _posts.Update(post);
         await _posts.Commit();
@@ -202,27 +216,40 @@ public class PostService : IPostService
         return post;
     }
 
-    public async Task<Post> UpdateContent(Uuid7 postId, Content content)
+    public async Task<Post> UpdateContent(Uuid7 postId, Uuid7 contentId, Content content)
     {
-        var post = await _posts.LookupPost(postId) 
+        var post = await _posts.LookupPost(postId)
                    ?? throw CoreException.MissingData<Post>("Can't find existing post to add content", postId);
         if (!post.FediId.HasLocalAuthority(_options))
             throw CoreException.WrongAuthority("Can't modify contents of remote post", post.FediId);
-        post.Contents.Remove(content);
-        post.Contents.Add(content);
-        
+        content.Id = contentId.ToGuid();
+        content.SetLocalFediId(_options);
+        if (post.Contents.FirstOrDefault(content) is not {} original)
+        {
+	        var details = new Dictionary<string, object>();
+	        details.Add("post", postId);
+	        details.Add("content", contentId);
+	        throw CoreException.InvalidRequest("Can't find content to update in post", details);
+	    }
+
+        original.UpdateFrom(content);
+        // _posts.Update(content);
+        // post.Contents.Remove(content);
+        // post.AddContent(content);
+        // post.Contents.Add(content);
+
         if (post.PublishedDate is not null)
         {
-            post.UpdatedDate = DateTimeOffset.Now;
+            post.UpdatedDate = DateTimeOffset.UtcNow;
             _postEvents.Published(post);
         }
-        
+
         _posts.Update(post);
         await _posts.Commit();
         _postEvents.Updated(post);
         return post;
     }
-    
+
     public async Task<Post> ReceiveCreate(Post post)
     {
         throw new NotImplementedException();
@@ -277,7 +304,7 @@ public class PostService : IPostService
     {
         throw new NotImplementedException();
     }
-    
+
     private async Task<Profile?> ResolveProfile(Uri profileId,
         Profile? onBehalfOf = null,
         [CallerMemberName] string name = "",
@@ -309,10 +336,10 @@ public class PostService : IPostService
         {
             _logger.LogError(ex, "Cannot resolve Profile {ProfileId}", profileId);
         }
-        
+
         return profile;
     }
-    
+
     private async Task<Post?> ResolvePost(Uri postId,
         Profile? onBehalfOf = null,
         [CallerMemberName] string name = "",
@@ -324,7 +351,7 @@ public class PostService : IPostService
         if (await _posts.LookupPost(postId) is { } post)
         {
             if (post.HasLocalAuthority(_options)) return post;
-            if (post.LastSeenDate >= DateTimeOffset.Now - TimeSpan.FromMinutes(10)) return post;
+            if (post.LastSeenDate >= DateTimeOffset.UtcNow - TimeSpan.FromMinutes(10)) return post;
             knownPost = true;
         }
 
@@ -346,5 +373,13 @@ public class PostService : IPostService
             _logger.LogWarning(e, "Cannot resolve post {Post}", postId);
             return default;
         }
+    }
+
+    public IAuthzPostService As(IEnumerable<Claim> claims, Uuid7 profileId)
+    {
+	    _profileId = profileId;
+	    _claims = claims;
+
+	    return this;
     }
 }
